+++ Objects/codeobject.c
@@ -1525,7 +1525,7 @@ _PyCode_GetNumFrameSlots(PyCodeObject *code)
 {
     /* This function needs to remain in sync with the calculation of
      * co_framesize  in Tools/build/deepfreeze.py */
-    assert(code->co_framesize >= (int)FRAME_SPECIALS_SIZE);
+    assert(code->co_framesize >= FRAME_SPECIALS_SIZE);
     return code->co_framesize - FRAME_SPECIALS_SIZE;
 }
